requirements.txt
request
bs4
selenium
pyperclip
pyautogui